Refactor: Refactoring main page
