+++ app/(backheaderRoute)/sign-up/_components/SignUpForm.tsx
@@ -5,7 +5,7 @@ import { ButtonSubmit } from "@/app/_components/ui/button/StyledButton";
 import { useForm } from "react-hook-form";
 import { useEffect, useState, useTransition } from "react";
 import { COLORS } from "@/public/styles/colors";
-import { signUp } from "@/app/api/auth/sign-up/route";
+import { signUp } from "@/public/utils/apis/signUp";
 import Loading from "@/app/_components/ui/loading/Loading";
 
 interface SignUpInputs {
