+++ app/(bigRollpeRouter)/rollpe/whole/[rollpeId]/page.tsx
@@ -2,7 +2,7 @@
 import styled from "styled-components";
 import { useParams, useRouter } from "next/navigation";
 import { useEffect, useState, useTransition } from "react";
-import { getRollpeDetail } from "@/app/api/rollpe/route";
+import { getRollpeDetail } from "@/public/utils/apis/rollpe";
 import Loading from "@/app/_components/ui/loading/Loading";
 import { Rollpe } from "@/public/utils/types";
 import HeartModal from "@/app/_components/ui/modal/HeartEditModal";
