+++ app/(logoMenuHeaderRoute)/mypage/my-rollpe/page.tsx
@@ -5,7 +5,7 @@ import { RollpeListItemProps } from "@/public/utils/types";
 import { RollpeList } from "@/app/_components/ui/list/RollpeList";
 import { Rollpe } from "@/public/utils/types";
 import { useEffect, useState, useTransition } from "react";
-import { getUserRollpeList } from "@/app/api/rollpe/route";
+import { getUserRollpeList } from "@/public/utils/apis/rollpe";
 import Loading from "@/app/_components/ui/loading/Loading";
 import { ButtonMore } from "@/app/_components/ui/button/StyledButton";
