+++ app/page.tsx
@@ -2,8 +2,8 @@
 import MainIntroSection from "./_components/sections/main/MainIntroSection";
 import MainRollpeGuideSection01 from "./_components/sections/main/MainRollpeGuideSection01";
 import MainRollpeGuideSection02 from "./_components/sections/main/MainRollpeGuideSection02";
-import MainRollpeGuideSection03 from "./_components/sections/main/MainRollpeGuidSection03";
 import styled from "styled-components";
+import Footer from "./_components/ui/layouts/Footer";
 
 const Home: React.FC = () => {
   return (
@@ -11,12 +11,12 @@ const Home: React.FC = () => {
       <MainIntroSection />
       <MainRollpeGuideSection01 />
       <MainRollpeGuideSection02 />
-      <MainRollpeGuideSection03 />
+      <Footer />
     </MainPageContentsContainer>
   );
 };
 
-const MainPageContentsContainer = styled.main`
+const MainPageContentsContainer = styled.section`
   display: flex;
   flex-direction: column;
   align-items: center;
